x86_64 build by top ./build.sh

+++ build_x86.sh
@@ -1,6 +1,6 @@
 #~/bin/bash
 if [ -z "$TARGET_ARCH" ];then
-export TARGET_ARCH=x86_64
+	export TARGET_ARCH=x86_64
 fi
 export ARCH=x86
 #check if build_x86 exists,
@@ -8,17 +8,32 @@ export ARCH=x86
 #mkdir ../build_$ARCH
 #cp -rf * ../build_x86
 #mv ../build_x86 .
-echo build_${TARGET_ARCH}
-if [ -d build_${TARGET_ARCH} ];then
-pushd build_${TARGET_ARCH}
-rm -rf *
-popd
-else
-mkdir -p build_${TARGET_ARCH}
+if [ -z "$FFTS_DIR" ]; then
+	export FFTS_DIR=`pwd`
 fi
-pushd build_${TARGET_ARCH}
+
+if [ -z "$FFTS_OUT" ]; then
+	echo build_${TARGET_ARCH}
+	if [ -d build_${TARGET_ARCH} ];then
+		pushd build_${TARGET_ARCH}
+		rm -rf *
+		popd
+	else
+		mkdir -p build_${TARGET_ARCH}
+	fi
+	export FFTS_OUT=build_${TARGET_ARCH}
+	local_build=1
+fi
+
+#if [ -f ${FFTS_OUT}/lib/libffts-${ARCH}.a ]; then
+rm -f ${FFTS_OUT}/lib/libffts-${ARCH}.a
+rm -f ${FFTS_OUT}/src/libffts.la
+#	rm -rf ${FFTS_OUT}/src/.libs
+#fi
+
 #clone the upper repo but discard .git
-git clone --depth=1 .. .
+git clone --depth=1 ${FFTS_DIR} ${FFTS_OUT}
+pushd ${FFTS_OUT}
 rm -rf .git .gitignore
 
 cp Makefile.am.${ARCH} Makefile.am
@@ -29,21 +44,21 @@ cp tests/Makefile.am.${ARCH} tests/Makefile.am
 automake --add-missing
 make
 
-popd
+if [ "$local_build" == "1" ]; then
+	popd
 
-#ln build_${TARGET_ARCH}/libffts-x86.a to lib
-if [ ! -d lib ];then
-	mkdir -p lib
-fi
+	#ln build_${TARGET_ARCH}/libffts-x86.a to lib
+	if [ ! -d lib ];then
+		mkdir -p lib
+	fi
 
-if [ -L lib/libffts-${ARCH}.a ];then
-	rm -f lib/libffts-${ARCH}.a
-fi
+	if [ -L lib/libffts-${TARGET_ARCH}.a ];then
+		rm -f lib/libffts-${TARGET_ARCH}.a
+	fi
 
-if [ -n "$FFTS_OUT" ]; then
-	rm -f ${FFTS_OUT}/libffts-${TARGET_ARCH}.a
-	cp `pwd`/build_${TARGET_ARCH}/lib/libffts-${ARCH}.a ${FFTS_OUT}/libffts-${TARGET_ARCH}.a
+	ln -s ${FFTS_OUT}/lib/libffts-${ARCH}.a lib/libffts-${TARGET_ARCH}.a
+
 else
 	rm -f lib/libffts-${TARGET_ARCH}.a
-	ln -s `pwd`/build_${TARGET_ARCH}/lib/libffts-${ARCH}.a lib/libffts-${TARGET_ARCH}.a
+	ln -s ${FFTS_OUT}/lib/libffts-${ARCH}.a lib/libffts-${TARGET_ARCH}.a
 fi
